requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
pyotp==2.8.0
qrcode[pil]==7.4.2
cryptography==41.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
